move binder to file
